blogo/back_end/pyproject.toml
[tool.poetry]
name = "django-simple-app"
version = "0.1.0"
description = ""
authors = ["Feyaaz <feyaazc94@gmail.com>"]
readme = "README.md"
packages = [{include = "django_simple_app"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.13"
Django = ">=3.2,<5.1"
strawberry-graphql-django = "^0.42.0"
strawberry-django-auth = "^0.376.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
